+++ metadata/__init__.py
@@ -1,6 +1,6 @@
 """Package metadata cache system."""
 
 from .metadata_cache import MetadataCacheService
-from .providers import MetadataProvider, WinGetProvider
+from .providers import MetadataProvider, WinGetProvider, ScoopProvider
 
-__all__ = ['MetadataCacheService', 'MetadataProvider', 'WinGetProvider']
+__all__ = ['MetadataCacheService', 'MetadataProvider', 'WinGetProvider', 'ScoopProvider']